backend/requirements.txt
# FocusFlow Backend Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# ML Dependencies
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
prophet==1.1.5

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Activity Tracker
pygetwindow==0.0.9
